requirements.txt
discord.py==2.5.2
langfuse==2.60.7
